Generate UML graph from plantuml

Generate an image containing based ion the plantuml code. TODO can I use vector formats?

# S3 method for plantuml
plot(x, file = NULL, width = 1024, height = NULL, css = NULL, ...)

Arguments

x

plantuml code to draw the UML graph

file

file is `NULL: the graph is drawn in the graphics device.

file is a file name: the graph is saved in the file and the type is based on the extensions.

file is NULL: the data which would have been saved in the file is returned in a character vector. This is only useful for text formats like eps or svg!

width

output width in pixels or NULL for default.

height

output height in pixels or NULL for default

css

path/url to external css file or raw vector with css data. This requires your system has a recent version of librsvg.

...

additional arguments for the plot function and the plantuml_run function.

Value

returns file name (including absolute path) of the created graph.

Examples

plantumlCode <- '
 @startuml
 (*) --> "First Activity"
 -->[You can put also labels] "Second Activity"
 --> (*)
 @enduml
'
if (FALSE) {
x <- as.plantuml( plantumlCode )
plot( x )
plot(as.plantuml(x), java_opt = "-Djava.awt.headless=true")
}