Run the plantuml binary

The in the package installation included plantuml binary is executed using the provided java and plantuml commands. This is effectively a wrapper around system2() with some values set to run plantuml.

plantuml_run(
  plantuml_jar = file.path(getPlantumlOption("jar_path"),
    getPlantumlOption("jar_name")),
  plantuml_opt = "-help",
  java_bin = getPlantumlOption("java_bin"),
  java_opt = getPlantumlOption("java_opt"),
  stdout = "",
  stderr = "",
  stdin = "",
  input = NULL,
  agree = FALSE
)

Arguments

plantuml_jar
path + filename to the plantuml jar file

plantuml_opt
options for plantuml. The default is -help to show all options of plantuml

java_bin
path to the java binary. The dafaulkt is java, i;e; it assumes that the binary is in the path

java_opt
options for the call of java. The default is -Djava.awt.headless=true -splash:no to enable a headless and silent execution of plantuml

stdout
See system2()

stderr
See system2()

stdin
See system2()

input
See system2()

agree
if TRUE, agree to create a directory into which to download the plantuml.jar file

Value

the result from the call to system2()

Examples

if (FALSE) {
# This will take some time when you run it
# for the first time as it will download \code{plantuml.jar}
  plantuml_run()
}