Run the plantuml binary

The in the package installation included plantuml binary is executed using the provided java and plantuml commands. This is effectively a wrapper around system2() with some values set to run plantuml.

plantuml_run(
  x = NULL,
  file = "",
  plantuml_jar = getPlantumlOption("jar_name"),
  plantuml_opt = getPlantumlOption("plantuml_opt"),
  java_bin = getPlantumlOption("java_bin"),
  java_opt = getPlantumlOption("java_opt"),
  wait = FALSE
)

Arguments

x

plantuml code to draw the UML graph

file

file name, including extension, to which the generated plantUML graph should be saved. The extension determines the format of the graph. If NULL', the graph is returned as a ASCII art, i.e. a character` vector..

plantuml_jar

path and name of the plantuml jar file. The dafault is read from getPlantumlOption("plantuml_jar").

plantuml_opt

options for the call of java. The default is read from getPlantumlOption("plantuml_opt").

java_bin

path to the java binary. The dafault is read from getPlantumlOption("java_bin").

java_opt

options for the call of java. The default is read from getPlantumlOption("java_opt").

wait

if TRUE, wait until the process has finished. If FALSE, return immediately.

Value

if wait is TRUE the pid of the process started. Otherwise the result from the call to system2()

Examples

if (FALSE) {
# This will take some time when you run it
# for the first time as it will download \code{plantuml.jar}
  plantuml_run()
}